refactor: テーマ案生成画面Container/Presentationalパターン廃止 #175
