--- front/src/app/[uuid]/generate-theme/layout.tsx
@@ -1,9 +1,0 @@
-import styles from "@/app/[uuid]/generate-theme/GenerateTheme.module.scss";
-
-export default function GenerateThemeLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <main className={styles.wrapper}>{children}</main>;
-}
