+++ front/src/features/generate-theme/components/AnswerForm/AnswerForm.tsx
@@ -9,33 +9,42 @@ import { AlertDialog, AlertDialogContent } from "@/components/ui/alert-dialog";
 import { LitUpBordersLg } from "@/components/ui/tailwind-buttons";
 import styles from "@/features/generate-theme/components/AnswerForm/AnswerForm.module.scss";
 import SelectOptions from "@/features/generate-theme/components/SelectOptions/SelectOptions";
+import { useUUIDCheck } from "@/hooks/useUUIDCheck";
 import { ThemeQuestionState, generateThemes } from "@/lib/actions";
 import { updateAIUsageHistory } from "@/lib/ai-usage-history";
 import { createIdeaSession, deleteIdeaSession } from "@/lib/idea-sessions";
 import { generateUUID } from "@/lib/uuid";
+import type { IdeaSessionType, Option } from "@/types";
 import { useRouter } from "next/navigation";
+import { useEffect, useState } from "react";
 import { useFormState, useFormStatus } from "react-dom";
 import toast from "react-hot-toast";
 import { BsExclamationTriangle } from "react-icons/bs";
 
 export default function AnswerForm({
-  isThemeGenerated,
-  retryCount,
-  setRetryCount,
-  isAlertModalOpen,
-  setIsAlertModalOpen,
-  uuid,
-  setHasApiError,
+  ideaSession,
+  aiGeneratedThemesArray,
 }: {
-  isThemeGenerated: boolean | undefined;
-  retryCount: number;
-  setRetryCount: React.Dispatch<React.SetStateAction<number>>;
-  isAlertModalOpen: boolean;
-  setIsAlertModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  uuid: string;
-  setHasApiError: React.Dispatch<React.SetStateAction<boolean>>;
+  ideaSession: IdeaSessionType | null;
+  aiGeneratedThemesArray: Option[];
 }) {
+  const [isThemeGenerated, setIsThemeGenerated] = useState(
+    ideaSession?.isAiThemeGenerated || false,
+  );
+  const [retryCount, setRetryCount] = useState(0);
+  const [isAlertModalOpen, setIsAlertModalOpen] = useState(false); // 無効な入力エラー画面の表示
+  const [hasApiError, setHasApiError] = useState(false); // APIリクエスト失敗時のエラーがあるかどうか
+  const [isMoveAlertModalOpen, setIsMoveAlertModalOpen] = useState(false); // AIアイデア生成済みエラー画面の表示
+
+  const { uuid } = useUUIDCheck({ ideaSession });
   const router = useRouter();
+
+  // テーマ生成ボタンの状態更新
+  useEffect(() => {
+    if (aiGeneratedThemesArray.length > 0) {
+      setIsThemeGenerated(true);
+    }
+  }, [aiGeneratedThemesArray]);
 
   // 質問、回答送信フォームの状態
   const initialQuestionState: ThemeQuestionState = {
@@ -49,7 +58,7 @@ export default function AnswerForm({
   >(async (prev: ThemeQuestionState | undefined, formData: FormData) => {
     const result = await generateThemes(prev, formData);
     if (result?.invalid) {
-      if (questionState?.errors?.option || questionState?.errors?.answer) {
+      if (result?.errors?.option || result?.errors?.answer) {
         toast.error("エラーがあるよ。確認してね。");
       }
       setIsAlertModalOpen(true);
@@ -85,6 +94,31 @@ export default function AnswerForm({
       await createIdeaSession(newUUID);
       router.push(`/${encodeURIComponent(newUUID)}/check-theme`);
     }
+  };
+
+  // APIリクエストエラー時のアラートで再試行をクリック時の処理
+  const handleRetryClick = () => {
+    setHasApiError(false);
+  };
+
+  // AIによるアイデア生成済みであれば、テーマ生成はせず、アイデア出し画面に遷移
+  useEffect(() => {
+    if (ideaSession?.isAiAnswerGenerated) {
+      setIsMoveAlertModalOpen(true);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  // 遷移先パスをプレフェッチ
+  useEffect(() => {
+    router.prefetch(`/${uuid}/generate-ideas`);
+  }, [uuid, router]);
+
+  // AIアイデア生成済みエラー画面でOKクリック時の処理
+  const handleMoveOkClick = () => {
+    setIsMoveAlertModalOpen(false);
+    router.push(`/${uuid}/generate-ideas`);
+    router.refresh();
   };
 
   return (
@@ -145,6 +179,26 @@ export default function AnswerForm({
             新しいセッションを作成するので、もう一度やってみてね。
           </>
         )}
+      </AlertModal>
+
+      {/* APIリクエストエラー時のアラート */}
+      <AlertModal
+        isOpen={hasApiError}
+        onClick={handleRetryClick}
+        actionDisplay="再試行"
+      >
+        ごめんね、AIの回答生成失敗です。再試行してみてね。
+      </AlertModal>
+
+      {/* すでにAIによるアイデア回答を生成済みの場合、アイデア出し画面に遷移するダイアログ表示 */}
+      <AlertModal
+        isOpen={isMoveAlertModalOpen}
+        onClick={handleMoveOkClick}
+        actionDisplay="OK"
+      >
+        このセッションで、すでにAIによるアイデア回答例を生成済みだよ。
+        <br />
+        アイデア出し画面に遷移するね。
       </AlertModal>
     </form>
   );
